deleted more unused files
