+++ src/js/containers/home/ProjectsManagementContainer.js
@@ -9,7 +9,6 @@ import OnlineImportModal from '../../components/home/projectsManagement/OnlineIm
 // actions
 import * as BodyUIActions from '../../actions/BodyUIActions';
 import * as MyProjectsActions from '../../actions/MyProjects/MyProjectsActions';
-import * as ImportOnlineActions from '../../actions/ImportOnlineActions';
 import * as ImportOnlineSearchActions from '../../actions/ImportOnlineSearchActions';
 import * as OnlineImportWorkflowActions from '../../actions/Import/OnlineImportWorkflowActions';
 import * as CSVExportActions from '../../actions/CSVExportActions';
@@ -106,7 +105,7 @@ const mapDispatchToProps = (dispatch) => {
         }));
       },
       handleURLInputChange: importLink => {
-        dispatch(ImportOnlineActions.getLink(importLink));
+        dispatch(OnlineImportWorkflowActions.getLink(importLink));
       },
       loadProjectFromLink: () => {
         dispatch(OnlineImportWorkflowActions.onlineImport());
